feat(home): Introduce new icons for AppBar actions
Add new vector drawables for the search (`search_line`) and menu (`menu_2_line`) actions.

Update the `HomeScreen`'s `AppBar` to use these new icons. The search icon is now an explicit search icon, and the settings icon has been replaced with a "more" menu icon, which navigates to the `MoreScreen`. The click handlers `onSettingsClick` and `onSettingsLongClick` have been renamed to `onMoreClick` and `onMoreLongClick` respectively to reflect this change.